requirements.txt
click==7.1.2
configparser==5.0.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
pkg-resources==0.0.0
pyserial==3.5
waitress==1.4.4
Werkzeug==1.0.1
